Alpha 0.02, WORKING. Refined input manager. Add model loading and camera. Moved test functionality in Testing. Added parametrics delegates. On the way to normal complex renderer.

+++ include/HarpyRavenPart/nest/windowing/window.h
@@ -151,7 +151,7 @@ namespace harpy::nest::windowing
         bool is_transparent();
         bool is_mouse_passthrough();
 
-        std::pair<int, int> get_screen_center();
+        std::pair<double, double> get_screen_center();
 
         ~window();
     };
