+++ src/HarpyRavenPart/nest/windowing/window.cpp
@@ -458,6 +458,6 @@ uint32_t harpy::nest::windowing::window::get_position_y() {
     return y;
 }
 
-std::pair<int, int> harpy::nest::windowing::window::get_screen_center() {
+std::pair<double, double> harpy::nest::windowing::window::get_screen_center() {
     return std::pair<int, int>(ci.width/2, ci.height/2);
 }
